Adds clippy to CI/CD for BSP and DLA (#39)
* Adds clippy to CI/CD
* Runs clippy for dla
* Fixes dla to pass tests
* Runs clippy for bsp
* Fixes bsp to to pass tests

+++ examples/headsail-bsp/src/uart.rs
@@ -10,5 +10,5 @@ pub fn uart_write(s: &str) {
 #[inline]
 pub fn putc(c: u8) {
     // Safety: we don't know if u8 writes work for all target architectures
-    unsafe { write_u8(UART0_ADDR, c as u8) };
+    unsafe { write_u8(UART0_ADDR, c) };
 }
